arild fixes relative joystick position

+++ adc.c
@@ -11,9 +11,9 @@ void adc_init( void )
     DDRB &= ~(1<<DDB3);
 }
 
-uint8_t adc_read(uint8_t channel)
+int16_t adc_read(uint8_t channel)
 {
     *ext_adc = channel;
     _delay_ms(40);
-    uint8_t adc_read = ext_adc[0];
+    uint16_t adc_read = ext_adc[0];
 }
